+++ doc/source/doxygen-docs/changelog.md
@@ -6,6 +6,8 @@
     - mrpt::opengl::CFBORender now does not rely on GLUT to create opengl contexts, but on EGL.
   - \ref mrpt_typemeta_grp
     - Add syntactic sugar function mrpt::typemeta::str2enum<>().
+- BUG FIXES:
+  - mrpt::opengl::CFBORender did only render the `main` viewport, it now processes all of them.
 
 # Version 2.4.1: Released Jan 5th, 2022
 - Changes in build system: