+++ doc/source/doxygen-docs/changelog.md
@@ -2,6 +2,8 @@
 
 # Version 2.4.2: UNRELEASED
 - Changes in libraries:
+  - \ref mrpt_containers_grp
+    - mrpt::container::yaml::operator(size_t) added, conditionally to `size_t` being a different type than `uint64_t` and such (Fixes build errors on OSX).
   - \ref mrpt_gui_grp
     - GUI windows can now have custom icons via mrpt::gui::CDisplayWindowGUI::setIcon() or mrpt::gui::CDisplayWindowGUI::setIconFromData()
   - \ref mrpt_img_grp